[BUGFIX] return float value from FloatTypeConverter

+++ Classes/TypeConverter/FloatTypeConverter.php
@@ -25,7 +25,7 @@ class FloatTypeConverter extends AbstractTypeConverter
      */
     public function convertFrom($sourceValue, int $qbankDataTypeId)
     {
-        return (int)$sourceValue;
+        return (float)$sourceValue;
     }
 
     /**
